Use single precision

+++ chvi/types.hpp
@@ -1,7 +1,7 @@
 #ifndef TYPES_HPP_
 #define TYPES_HPP_
 
-typedef double coordinate;
+typedef float coordinate;
 
 #include <vector>
 typedef std::vector<coordinate> point;
